reafactor: clean code

+++ src/main/java/com/example/slotify_backend/service/client/ClientPanelService.java
@@ -5,6 +5,7 @@ import com.example.slotify_backend.dto.client.ClientDetailsDTO;
 import com.example.slotify_backend.entity.Client;
 import com.example.slotify_backend.entity.Event;
 import com.example.slotify_backend.entity.User;
+import com.example.slotify_backend.exception.UserNotFoundException;
 import com.example.slotify_backend.mapper.ClientDetailsMapper;
 import com.example.slotify_backend.repository.ClientRepository;
 import com.example.slotify_backend.repository.EventRepository;
@@ -13,6 +14,7 @@ import com.example.slotify_backend.service.company.JwtService;
 import com.example.slotify_backend.service.company.SupabaseStorageService;
 import jakarta.transaction.Transactional;
 import lombok.AllArgsConstructor;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -31,15 +33,15 @@ public class ClientPanelService {
     private final ClientRepository clientRepository;
 
     public ClientDetailsDTO getClientDetails(String authHeader) {
-        Long clientId = jwtService.getUserIdFromAuthHeader(authHeader);
-        User client = userRepository.findById(clientId).orElseThrow(() -> new RuntimeException("User not found") );
+        Long clientId = getUserFromAuthHeader(authHeader);
+        User client = getUserOrThrow(clientId);
 
         return clientDetailsMapper.toDTO(client);
     }
 
     public void updateClientDetails(ClientDetailsDTO clientDetailsDTO, String authHeader) {
-        Long clientId = jwtService.getUserIdFromAuthHeader(authHeader);
-        User userAccount = userRepository.findById(clientId).orElseThrow(() -> new RuntimeException("User not found"));
+        Long clientId = getUserFromAuthHeader(authHeader);
+        User userAccount = getUserOrThrow(clientId);
         Client clientAccount = clientRepository.findByUserAccountId(clientId);
 
 
@@ -49,10 +51,10 @@ public class ClientPanelService {
 
     @Transactional
     public void uploadPictures(MultipartFile clientPic, String authHeader){
-        Long userId = jwtService.getUserIdFromAuthHeader(authHeader);
-        User client = userRepository.findById(userId).orElseThrow(() -> new RuntimeException("User not found"));
+        Long userId = getUserFromAuthHeader(authHeader);
+        User client = getUserOrThrow(userId);
 
-            if(clientPic != null){
+            if(clientPic != null && !clientPic.isEmpty()){
                 if(client.getPictureURL() != null){
                     supabaseStorageService.deletePictureByPublicUrl(client.getPictureURL());
                 }
@@ -65,7 +67,6 @@ public class ClientPanelService {
         Long userId = jwtService.getUserIdFromAuthHeader(authHeader);
         Client client = clientRepository.findByUserAccountId(userId);
 
-
         List<Event> events = eventRepository.findAllByClientId(client.getId());
         List<ClientBookingsDTO> clientBookingsDTO = new ArrayList<>();
         for (Event event : events) {
@@ -73,5 +74,13 @@ public class ClientPanelService {
         }
         return clientBookingsDTO;
 
-    };
+    }
+
+    private User getUserOrThrow(Long userId){
+        return userRepository.findById(userId).orElseThrow(() -> new UserNotFoundException("User not found"));
+    }
+
+    private Long getUserFromAuthHeader(String authHeader){
+        return jwtService.getUserIdFromAuthHeader(authHeader);
+    }
 }
